[NTPII-22] Integracja przycisku ze strona zwrotu samochodu
